# add netipds for comparison