Hi @gaissmai, I started working on [netipds](https://github.com/aromatt/netipds) in December 2023, shortly before you open-sourced bart. I stumbled upon iprbench and was very curious to find out how netipds would perform. I found that with some easy optimizations, netipds was competitive with most of the libraries compared here. It's still a work in progress, but I'd love to add it to your benchmarks!

Here are the results I got for bart vs. netipds.

```
goos: linux
goarch: amd64
cpu: AMD EPYC 7R32
                       │ bart/size.bm │           netipds/size.bm           │
                       │ bytes/route  │ bytes/route  vs base                │
Tier1PfxSize/1_000         106.9 ± 2%    100.5 ± 2%    -5.99% (p=0.002 n=6)
Tier1PfxSize/10_000        79.54 ± 0%    96.05 ± 0%   +20.76% (p=0.002 n=6)
Tier1PfxSize/100_000       35.73 ± 0%    94.33 ± 0%  +164.01% (p=0.002 n=6)
Tier1PfxSize/200_000       24.14 ± 0%    93.27 ± 0%  +286.37% (p=0.002 n=6)
RandomPfx4Size/1_000      101.60 ± 2%    99.65 ± 2%    -1.92% (p=0.002 n=6)
RandomPfx4Size/10_000      75.59 ± 0%    94.17 ± 0%   +24.58% (p=0.002 n=6)
RandomPfx4Size/100_000     81.65 ± 0%    86.53 ± 0%    +5.98% (p=0.002 n=6)
RandomPfx4Size/200_000     72.96 ± 0%    82.90 ± 0%   +13.62% (p=0.002 n=6)
RandomPfx6Size/1_000      105.80 ± 2%    99.44 ± 2%    -6.01% (p=0.002 n=6)
RandomPfx6Size/10_000      85.70 ± 0%    95.17 ± 0%   +11.05% (p=0.002 n=6)
RandomPfx6Size/100_000    106.10 ± 0%    92.67 ± 0%   -12.66% (p=0.002 n=6)
RandomPfx6Size/200_000    101.00 ± 0%    91.83 ± 0%    -9.08% (p=0.002 n=6)
RandomPfxSize/1_000       104.20 ± 2%    99.46 ± 2%    -4.55% (p=0.002 n=6)
RandomPfxSize/10_000       75.58 ± 0%    94.28 ± 0%   +24.74% (p=0.002 n=6)
RandomPfxSize/100_000      82.47 ± 0%    87.83 ± 0%    +6.50% (p=0.002 n=6)
RandomPfxSize/200_000      78.36 ± 0%    84.80 ± 0%    +8.22% (p=0.002 n=6)
geomean                    77.40         93.15        +20.35%
```

```
goos: linux
goarch: amd64
cpu: AMD EPYC 7R32
                                     │ bart/lpm.bm  │             netipds/lpm.bm              │
                                     │    sec/op    │     sec/op      vs base                 │
LpmTier1Pfxs/RandomMatchIP4            25.66n ±  7%    39.95n ±   7%   +55.68% (p=0.000 n=20)
LpmTier1Pfxs/RandomMatchIP6            23.09n ± 21%    31.46n ±  28%   +36.28% (p=0.004 n=20)
LpmTier1Pfxs/RandomMissIP4             24.43n ± 50%    52.69n ±   7%  +115.72% (p=0.000 n=20)
LpmTier1Pfxs/RandomMissIP6             5.537n ±  0%   18.205n ± 110%  +228.79% (p=0.000 n=20)
LpmRandomPfxs/1_000/RandomMatchIP4     7.361n ±  0%   44.290n ±   5%  +501.73% (p=0.000 n=20)
LpmRandomPfxs/1_000/RandomMatchIP6     15.99n ±  6%    18.97n ±   0%   +18.60% (p=0.000 n=20)
LpmRandomPfxs/1_000/RandomMissIP4      17.98n ± 36%    45.22n ±   4%  +151.43% (p=0.000 n=20)
LpmRandomPfxs/1_000/RandomMissIP6      12.67n ± 56%    21.76n ±   6%   +71.81% (p=0.000 n=20)
LpmRandomPfxs/10_000/RandomMatchIP4    21.45n ±  9%    51.73n ±  14%  +141.17% (p=0.000 n=20)
LpmRandomPfxs/10_000/RandomMatchIP6    20.54n ±  6%    24.93n ±  12%   +21.37% (p=0.003 n=20)
LpmRandomPfxs/10_000/RandomMissIP4     24.50n ±  4%    59.64n ±   1%  +143.40% (p=0.000 n=20)
LpmRandomPfxs/10_000/RandomMissIP6     20.19n ± 14%    28.07n ±   4%   +39.03% (p=0.000 n=20)
LpmRandomPfxs/100_000/RandomMatchIP4   17.65n ± 46%    27.62n ±   0%   +56.49% (p=0.000 n=20)
LpmRandomPfxs/100_000/RandomMatchIP6   16.52n ± 48%    18.23n ±  23%         ~ (p=0.409 n=20)
LpmRandomPfxs/100_000/RandomMissIP4    35.50n ±  4%    27.61n ±   0%   -22.24% (p=0.000 n=20)
LpmRandomPfxs/100_000/RandomMissIP6    26.11n ±  5%    33.65n ±   3%   +28.90% (p=0.000 n=20)
geomean                                18.05n          31.64n          +75.30%
```
```
goos: linux
goarch: amd64
cpu: AMD EPYC 7R32
                         │ bart/update.bm │           netipds/update.bm           │
                         │   sec/route    │  sec/route    vs base                 │
InsertRandomPfxs/1_000        37.39n ± 0%   180.20n ± 0%   +381.95% (p=0.002 n=6)
InsertRandomPfxs/10_000       50.37n ± 1%   266.95n ± 0%   +429.98% (p=0.002 n=6)
InsertRandomPfxs/100_000      76.38n ± 0%   365.65n ± 0%   +378.72% (p=0.002 n=6)
InsertRandomPfxs/200_000      134.4n ± 1%    450.8n ± 0%   +235.58% (p=0.002 n=6)
DeleteRandomPfxs/1_000        16.77n ± 1%   116.60n ± 1%   +595.29% (p=0.002 n=6)
DeleteRandomPfxs/10_000       16.50n ± 1%   182.05n ± 1%  +1003.33% (p=0.002 n=6)
DeleteRandomPfxs/100_000      16.33n ± 0%   288.30n ± 0%  +1666.00% (p=0.002 n=6)
DeleteRandomPfxs/200_000      16.62n ± 1%   366.75n ± 0%  +2106.68% (p=0.002 n=6)
geomean                       33.13n         254.9n        +669.29%
```