@gaissmai 2025-01-16
Hi @aromatt thanks for the PR, I'll merge it, cuerrently I'm doing the benchstat statistics for netipds on my laptop, after that I'll merge and change the README with the link to your library and benchstat values.