@gaissmai 2025-01-16
your library looks promising, I'll look over it the next days, maybe I learn something new or can give a tip to you. That's open source, I like it!